Layout for GUI window displaying processed image
Created a non-functional GUI window to display the processed image and
save the image in a user selected folder location as a user selected
file type. Non-functional buttons created for opening comparison and
histogram windows. The requirements.txt file was also updated to include
the Pillow library for image display.